Implement grayscale morphological reconstruction
topotoolbox fills sinks in digital elevation models using grayscale
morphological reconstruction. This algorithm repeatedly dilates a
"marker" image while constraining it to lie below a "mask" image. The
negative DEM is used as the mask while the marker is identical to the
negative DEM on the boundary pixels but negative infinity in its
interior. The morphological reconstruction iteratively fills in the
interior of the marker image, but does not descend into sinks of the
mask.

src/morphology/reconstruct.c implements the sequential reconstruction
algorithm presented by Vincent (1993). The algorithm repeatedly scans
in alternating directions through the DEM, updates the marker image
based on the values of the marker and mask in a neighborhood around
the given pixel and terminates when the marker stops changing. The
sequential reconstruction algorithm is not as efficient as a hybrid
algorithm presented by Vincent that combines sequential reconstruction
with a breadth-first search. However, the hybrid algorithm requires
managing a queue of pixels that remain to be updated, while the
sequential algorithm requires no memory allocation beyond the marker
and mask images.

src/morphology/reconstruct.h exposes the reconstruct function for use
elsewhere in topotoolbox. This private header is not, however, exposed
to users of the library.