Implement sink filling using morphological reconstruction
src/fillsinks.c implements `fillsinks` using `reconstruct`. The input
and output digital elevation models should be passed in by the caller
as pointers to arrays of floats. The input DEM is temporarily
modified, but then reverted to its original state before the function
returns. This should be safe as long as concurrent accesses to the DEM
are avoided, but care should be taken in the future.

include/topotoolbox.h is modified to expose the `fillsinks` function
to users of the library.