Add a .clang-format file and run format checks during CI
This commit begins to solidify a style guide for
libtopotoolbox (#13). clang-format addresses code formatting and not
other aspects of style, however, which will need to be managed using
other tools.

.clang-format is the clang-format configuration file output by
clang-format -style=Google --dump-config corresponding to the Google
C++ style guide

.github/workflows/ci.yaml is changed to use the
jidicula/clang-format-action to run clang-format on the source files
in the repository.