Only run AddressSanitizer on GCC/clang compilers
There is a problem linking to the ASan libraries on Windows on GitHub
Actions. This commit leaves these lines commented out so they can be
added back in if the linking problem can be resolved and we run into
errors that on Windows that ASan does not catch on other platforms.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>