Undefine NDEBUG flag
This allows us to use the `assert` macro in our tests in Release
mode. Otherwise the asserts are no-ops, and the tests are
useless. Since the flag only applies to the `random_dem` tests, it
should not affect any asserts in the compiled library.

Eventually we may want to use a more complete testing framework with
proper testing macros.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>